Backend/requirement.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.4.2
motor>=3.3.2
pymongo>=4.6.0
beautifulsoup4>=4.12.2
httpx>=0.25.2
lxml>=4.9.3
chardet>=5.2.0
torch>=2.1.0
transformers>=4.36.0
scikit-learn>=1.3.2
numpy>=1.26.0
